+++ moq-karp/src/video/dimensions.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 pub struct Dimensions {
-	pub width: u16,
-	pub height: u16,
+	pub width: u32,
+	pub height: u32,
 }
